Cleaned miner packaged. This commit removes most of the unwanted constructors becoming to the go-ethereum project, leaving a simple base for future work specific to Ellcrys.
