+++ miner/miner_suite_test.go
@@ -8,7 +8,6 @@ import (
 )
 
 func TestMiner(t *testing.T) {
-	log.SetToDebug()
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Miner Suite")
 }
